requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
numpy==1.21.2
pandas==1.3.2
pydantic==1.8.2
PyJWT~=2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv~=0.20.0
pytz==2021.1
requests~=2.27.1
six==1.16.0
SQLAlchemy~=1.4.36
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
matplotlib~=3.6.0
pyinstaller==5.7.0